requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-google-genai>=0.0.6
langchain-community>=0.0.20
langgraph>=0.0.20
python-dotenv>=1.0.0
pandas>=2.0.0
requests>=2.31.0
openai>=1.10.0
google-generativeai>=0.3.0

